Refactor tool_use stuff
